refactors and adding tests

+++ libdrbug/src/process.rs
@@ -29,21 +29,49 @@ pub enum ProcessState {
     Terminated,
 }
 
+#[derive(Debug, Default)]
+struct WaitOptions {
+    terminate_on_end: bool,
+    attached: bool,
+}
+
 #[derive(Debug)]
 pub struct Process {
     pid: Pid,
     state: ProcessState,
-    terminate_on_end: bool,
+    opts: WaitOptions,
 }
 
 impl Process {
     pub fn attach(pid_int: i32) -> anyhow::Result<Self> {
         let pid = Pid::from_raw(pid_int);
         ptrace::attach(pid)?;
-        Self::wait_new(pid, false)
+        Self::new_then_wait(pid, WaitOptions { terminate_on_end: false, attached: true })
     }
 
-    pub fn launch(path: String) -> anyhow::Result<Self> {
+    pub fn launch(path: &str) -> anyhow::Result<Self> {
+        Self::launch_maybe_attach(path, true)
+    }
+
+    pub fn resume(&mut self) -> Empty {
+        ptrace::cont(self.pid, None)?;
+        self.state = ProcessState::Running;
+        Ok(())
+    }
+
+    pub fn wait_on_signal(&self) -> anyhow::Result<WaitStatus> {
+        waitpid(self.pid, None).map_err(|e| e.into())
+    }
+
+    pub fn pid(&self) -> Pid {
+        self.pid
+    }
+
+    pub fn state(&self) -> ProcessState {
+        self.state
+    }
+
+    fn launch_maybe_attach(path: &str, attach: bool) -> anyhow::Result<Self> {
         let mut channel = Pipe::new_exec_safe()?;
 
         let path_cstring = CString::new(path)?;
@@ -55,10 +83,17 @@ impl Process {
             // in the child process, we can't just use `?`, since it won't get
             // communicated back to the parent; instead we use the channel.
             // Then we still return the error so the child process exits.
-            if let Err(e) = ptrace::traceme() {
-                let _ = write!(&mut channel, "tracing failed: {e:?}");
-                return Err(e.into());
+
+            if attach {
+                // ptrace::traceme sets up the "attach" in reverse, the child sets up tracing, and
+                // then Linux will pause the process on an exec call; this is why we don't need an
+                // explicit ptrace::attach call in this case
+                if let Err(e) = ptrace::traceme() {
+                    let _ = write!(&mut channel, "tracing failed: {e:?}");
+                    return Err(e.into());
+                }
             }
+
             // execvp returns Result<Infallible>, i.e., if it ever returns anything
             // it is guaranteed to be an Error, hence the irrefutable_let_patterns
             // warning here; however, I still think it's more clear to use the if let,
@@ -103,38 +138,18 @@ impl Process {
         channel.close_reader();
 
         if !data.is_empty() {
-            println!("{data:?}");
             waitpid(child, None)?;
             bail!("child process failed with error: {}", std::str::from_utf8(&data)?);
         }
-        Self::wait_new(child, true)
+
+        Self::new_then_wait(child, WaitOptions { terminate_on_end: true, attached: attach })
     }
 
-    pub fn resume(&mut self) -> Empty {
-        ptrace::cont(self.pid, None)?;
-        self.state = ProcessState::Running;
-        Ok(())
-    }
-
-    pub fn wait_on_signal(&self) -> anyhow::Result<WaitStatus> {
-        waitpid(self.pid, None).map_err(|e| e.into())
-    }
-
-    pub fn pid(&self) -> Pid {
-        self.pid
-    }
-
-    pub fn state(&self) -> ProcessState {
-        self.state
-    }
-
-    fn wait_new(pid: Pid, terminate_on_end: bool) -> anyhow::Result<Self> {
-        let proc = Process {
-            pid,
-            terminate_on_end,
-            state: ProcessState::Stopped,
-        };
-        proc.wait_on_signal()?;
+    fn new_then_wait(pid: Pid, opts: WaitOptions) -> anyhow::Result<Self> {
+        let proc = Process { pid, state: ProcessState::Stopped, opts };
+        if proc.opts.attached {
+            proc.wait_on_signal()?;
+        }
         Ok(proc)
     }
 }
@@ -142,18 +157,27 @@ impl Process {
 impl Drop for Process {
     // No error handling; "You'll just have to believe in your destructor"
     fn drop(&mut self) {
-        // Must be stopped before you can ptrace::detach
-        if self.state == ProcessState::Running {
-            let _ = kill(self.pid, Signal::SIGSTOP);
-            let _ = waitpid(self.pid, None);
+        if self.opts.attached {
+            // Must be stopped before you can ptrace::detach
+            if self.state == ProcessState::Running {
+                let _ = kill(self.pid, Signal::SIGSTOP);
+                let _ = waitpid(self.pid, None);
+            }
         }
 
         let _ = ptrace::detach(self.pid, None);
         let _ = kill(self.pid, Signal::SIGCONT);
 
-        if self.terminate_on_end {
+        if self.opts.terminate_on_end {
             let _ = kill(self.pid, Signal::SIGKILL);
             let _ = waitpid(self.pid, None);
         }
+    }
+}
+
+#[cfg(test)]
+impl Process {
+    pub(crate) fn launch_no_attach(path: &str) -> anyhow::Result<Self> {
+        Self::launch_maybe_attach(path, false)
     }
 }
